代码优化

+++ tcc-transaction-dashboard/src/main/java/org/mengyun/tcctransaction/dashboard/controller/DomainController.java
@@ -4,11 +4,13 @@ import org.mengyun.tcctransaction.dashboard.constants.DashboardConstant;
 import org.mengyun.tcctransaction.dashboard.dto.DomainStoreDto;
 import org.mengyun.tcctransaction.dashboard.dto.DomainStoreRequestDto;
 import org.mengyun.tcctransaction.dashboard.dto.ResponseDto;
+import org.mengyun.tcctransaction.dashboard.enums.ConnectionMode;
 import org.mengyun.tcctransaction.dashboard.service.DomainService;
 import org.mengyun.tcctransaction.dashboard.service.TaskService;
 import org.mengyun.tcctransaction.utils.AlertUtils;
 import org.mengyun.tcctransaction.utils.CollectionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -27,6 +29,9 @@ public class DomainController {
 
     @Autowired
     private TaskService taskService;
+
+    @Value("${spring.tcc.dashboard.connection-mode:embedded}")
+    private String connectionMode;
 
     @GetMapping("/allKeys")
     @ResponseBody
@@ -55,9 +60,12 @@ public class DomainController {
     @RequestMapping("/delete")
     @ResponseBody
     public ResponseDto delete(@RequestBody DomainStoreRequestDto requestDto) {
-        ResponseDto taskDeleteResponseDto = taskService.delete(requestDto.getDomain());
-        if (!taskDeleteResponseDto.isSuccess()) {
-            return taskDeleteResponseDto;
+        // server模式时，删除domain前先删除任务
+        if(ConnectionMode.SERVER.name().equals(connectionMode.toUpperCase())){
+            ResponseDto taskDeleteResponseDto = taskService.delete(requestDto.getDomain());
+            if (!taskDeleteResponseDto.isSuccess()) {
+                return taskDeleteResponseDto;
+            }
         }
         return domainService.delete(requestDto);
     }
